+++ OLCore/Classes/Extensions/UIImage+Crop.swift
@@ -50,4 +50,34 @@ extension UIImage {
         UIGraphicsEndImageContext()
         return scaledImage ?? self
     }
+
+    public func cropToSize(newSize: CGSize) -> UIImage {
+        let size = self.size
+        var x: CGFloat!
+        var y: CGFloat!
+        x = (size.width - newSize.width) * 0.5
+        y = (size.height - newSize.height) * 0.5
+        var tempNewSize = newSize
+        if (self.imageOrientation == .left || self.imageOrientation == .leftMirrored || self.imageOrientation == .right || self.imageOrientation == .rightMirrored) {
+            var temp = x
+            x = y
+            y = temp
+            
+            temp = tempNewSize.width
+            tempNewSize.width = tempNewSize.height
+            tempNewSize.height = temp!
+        }
+        let cropRect = CGRect(x: x * self.scale, y: y * self.scale, width: tempNewSize.width * self.scale, height: tempNewSize.height * self.scale)
+        let croppedImageRef = self.cgImage!.cropping(to: cropRect)
+        let cropped = UIImage(cgImage: croppedImageRef!, scale: self.scale, orientation: self.imageOrientation)
+        return cropped
+    }
+
+    public func fixImageOrientation() -> UIImage {
+        UIGraphicsBeginImageContext(self.size)
+        self.draw(at: .zero)
+        let newImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return newImage ?? self
+    }
 }
